🧹 Extract matrix minigame logic to hooks
