+++ src/components/effects/Matrix/AuthContext.tsx
@@ -250,7 +250,7 @@ export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
     [DEVICE_KEYS.MOBILE]: isMobileUnlocked,
   } = unlockState;
 
-  // biome-ignore lint/correctness/useExhaustiveDependencies: Explicit dependency management
+
   const toolsAccessible = useMemo(() => {
     if (isMobile) {
       return isMobileUnlocked;
@@ -260,7 +260,7 @@ export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
 
   return (
     <AuthContext.Provider
-      // biome-ignore lint/correctness/useExhaustiveDependencies: Explicit dependency management
+
       value={useMemo(
         () => ({
           isUnlocked,
